Adjusting rates for consistency between runs
Setting the search volume coefficient to `4000 * temp_eff` so that the predation and encounter rates depend only on temperature, not calibrated by mizer.
Metabolism and external mortality both set to zero. These may be added back in later. External mortality for this model should not be set with the `ext_mort_prop` option but with `setExtMort`.